changes into ProjectMapper and add LowerCase for Tags
